refactor(04-02): extract hooks concern to shared/concerns/hooks.ts
Move appendHooksBlock and the hook summary types (ClaudeHookEvent,
_ToolEvent, HookSummaryEntry, HookSummary) plus their import-fence
rationale out of notify.ts into a new shared/concerns/hooks.ts module
(MOD-04, D-01 direct-function-call wiring). Per the cleaner cut,
COMPONENT_KINDS and appendResolvedComponentLines stay in notify.ts; only
the kind === 'hooks' arm dispatches into the concern, and the call-site
appendHooksBlock(lines, components.hooks) stays byte-identical.

The concern imports nothing from notify.ts (it owns its types). The
shared/->domain/ fence is preserved: the types stay on the shared/ side
so domain/components/hook-events.ts keeps its downward satisfies-pin.
Hook-type importers (info.ts, hook-events.ts, notify-v2.test.ts) are
repointed to the concern. Output-neutral: catalog-uat 4/4 green and
docs/output-catalog.md byte-unchanged.
